-Added some security checks -Added the actual eventos_edit.php that shows up on clicking "Eventos" if logged in -Added Login/Logoff logic -Finally a button to delete the selected event -ordered events by adc chronology:wq:wq:wq

+++ src/views/admin/eventos_del.php
@@ -1,6 +1,11 @@
 <?php
-require("../../Classes/Evento.php");
+require_once("../../config.php");
+if (empty($_SESSION['admin'])) {
+    http_response_code(403);
+    exit;
+}
+
 $Eventos = new Evento("../../data/eventos.json");
 $Eventos->delete($_POST["id"]);
-header("Location: /views/admin/eventos_edit.php");
+header("Location: /index.php?page=admin/eventos_edit");
 ?>
